Desktop/RAG QnA/requirements.txt
llama
torch
langchain
sentence-transformers
chromadb
pymupdf
llama-cpp-python